# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: BaCS API
  description: API сервиса по бронированию рабочих мест в коворкингах BaCS.
  version: v1
paths:
  /locations:
    get:
      tags:
        - Locations
      summary: Получить список локаций.
      parameters:
        - name: ids
          in: query
          description: Фильтр по ID резерваций.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID резерваций.
            nullable: true
        - name: offset
          in: query
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: limit
          in: query
          schema:
            maximum: 100
            minimum: 0
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDtoPaginatedResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - Locations
      summary: Создать локацию.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/locations/{locationId}':
    get:
      tags:
        - Locations
      summary: Получить локацию по ID.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Locations
      summary: Обновить локацию.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Locations
      summary: Удалить локацию.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/locations/{locationId}/image':
    put:
      tags:
        - Locations
      summary: Загрузить фотографию локации.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  description: Фотография локации.
                  format: binary
            encoding:
              file:
                style: form
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Locations
      summary: Удалить фотографию локации.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/locations/{locationId}/admins/{userId}':
    put:
      tags:
        - Locations
      summary: Добавить администратора локации.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
        - name: userId
          in: path
          description: ID пользователя.
          required: true
          schema:
            type: string
            description: ID пользователя.
            format: uuid
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Locations
      summary: Удалить администратора локации.
      parameters:
        - name: locationId
          in: path
          description: ID локации.
          required: true
          schema:
            type: string
            description: ID локации.
            format: uuid
        - name: userId
          in: path
          description: ID пользователя.
          required: true
          schema:
            type: string
            description: ID пользователя.
            format: uuid
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /reservations:
    get:
      tags:
        - Reservations
      summary: Получить список резерваций.
      parameters:
        - name: ids
          in: query
          description: Фильтр по ID резерваций.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID резерваций.
            nullable: true
        - name: userIds
          in: query
          description: Фильтр по ID пользователей.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID пользователей.
            nullable: true
        - name: locationIds
          in: query
          description: Фильтр по ID локаций.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID локаций.
            nullable: true
        - name: resourceIds
          in: query
          description: Фильтр по ID ресурсов.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID ресурсов.
            nullable: true
        - name: offset
          in: query
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: limit
          in: query
          schema:
            maximum: 100
            minimum: 0
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDtoPaginatedResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - Reservations
      summary: Создать резервацию.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/reservations/{reservationId}':
    get:
      tags:
        - Reservations
      summary: Получить резервацию по ID.
      parameters:
        - name: reservationId
          in: path
          description: ID резервации.
          required: true
          schema:
            type: string
            description: ID резервации.
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Reservations
      summary: Обновить резервацию.
      parameters:
        - name: reservationId
          in: path
          description: ID резервации.
          required: true
          schema:
            type: string
            description: ID резервации.
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/reservations/{reservationId}/Cancel':
    put:
      tags:
        - Reservations
      summary: Отменить резервацию.
      parameters:
        - name: reservationId
          in: path
          description: ID резервации.
          required: true
          schema:
            type: string
            description: ID резервации.
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
    CalendarSettingsDto:
      type: object
      properties:
        availableFrom:
          type: string
          description: Время начала работы локации.
          format: time
        availableTo:
          type: string
          description: Время окончания работы локации.
          format: time
        availableDaysOfWeek:
          type: array
          items:
            $ref: '#/components/schemas/RussianDayOfWeek'
          description: Набор рабочих дней недели локации.
          nullable: true
      additionalProperties: false
    CreateLocationRequest:
      required:
        - address
        - calendarSettings
        - name
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: Название локации.
        address:
          minLength: 1
          type: string
          description: Адресс локации.
        description:
          type: string
          description: Описание локации.
          nullable: true
        calendarSettings:
          $ref: '#/components/schemas/CalendarSettingsDto'
      additionalProperties: false
    CreateReservationRequest:
      required:
        - from
        - locationId
        - resourceId
        - to
      type: object
      properties:
        resourceId:
          minLength: 1
          type: string
          description: ID ресурса для бронирования.
          format: uuid
        locationId:
          minLength: 1
          type: string
          description: 'ID локации, в которой находится ресурс.'
          format: uuid
        from:
          minLength: 1
          type: string
          description: Дата и время начала бронирования.
          format: date-time
        to:
          minLength: 1
          type: string
          description: Дата и время окончания бронирования.
          format: date-time
      additionalProperties: false
...
    LocationDto:
      type: object
      properties:
        id:
          type: string
          description: ID локации.
          format: uuid
        name:
          type: string
          description: Название локации.
          nullable: true
        address:
          type: string
          description: Адрес локации.
          nullable: true
        description:
          type: string
          description: Описание локации.
          nullable: true
        imageUrl:
          type: string
          description: Ссылка на фотографию локации.
          nullable: true
        calendarSettings:
          $ref: '#/components/schemas/CalendarSettingsDto'
      additionalProperties: false
    LocationDtoPaginatedResponse:
      type: object
      properties:
        collection:
          type: array
          items:
            $ref: '#/components/schemas/LocationDto'
          nullable: true
        totalCount:
          type: integer
          format: int32
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    ReservationDto:
      type: object
      properties:
        id:
          type: string
          description: ID резервации.
          format: uuid
        resourceId:
          type: string
          description: ID забронированного ресурса.
          format: uuid
        locationId:
          type: string
          description: 'ID локации, в которой забронирован ресурс.'
          format: uuid
        userId:
          type: string
          description: 'ID пользователя, забронировавшего ресурс.'
          format: uuid
        from:
          type: string
          description: Дата и время начала резервации.
          format: date-time
        to:
          type: string
          description: Дата и время окончания резервации.
          format: date-time
        status:
          $ref: '#/components/schemas/ReservationStatus'
      additionalProperties: false
    ReservationDtoPaginatedResponse:
      type: object
      properties:
        collection:
          type: array
          items:
            $ref: '#/components/schemas/ReservationDto'
          nullable: true
        totalCount:
          type: integer
          format: int32
      additionalProperties: false
    ReservationStatus:
      enum:
        - Created
        - Cancelled
      type: string
...
    RussianDayOfWeek:
      enum:
        - Monday
        - Tuesday
        - Wednesday
        - Thursday
        - Friday
        - Saturday
        - Sunday
      type: string
    UpdateLocationRequest:
      required:
        - address
        - calendarSettings
        - name
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: Название локации.
        address:
          minLength: 1
          type: string
          description: Адрес локации.
        description:
          type: string
          description: Описание локации.
          nullable: true
        calendarSettings:
          $ref: '#/components/schemas/CalendarSettingsDto'
      additionalProperties: false
    UpdateReservationRequest:
      required:
        - from
        - to
      type: object
      properties:
        from:
          minLength: 1
          type: string
          description: Время начала резервации.
          format: date-time
        to:
          minLength: 1
          type: string
          description: Время окончания резервации.
          format: date-time
      additionalProperties: false
...
  securitySchemes:
    Keycloak:
      type: oauth2
      description: Аутентификация через Keycloak
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:7080/keycloak/realms/bacs/protocol/openid-connect/auth
          tokenUrl: http://localhost:7080/keycloak/realms/bacs/protocol/openid-connect/token
          scopes:
            bacs-api: 'Обязательный скоуп, необходимый для авторизации на BaCS API'
          x-usePkce: SHA-256
security:
  - Keycloak: [ ]